Refactor YesNoButton
