+++ src/frontend/lib/components/YesNoButton.tsx
@@ -1,7 +1,7 @@
 import { t } from '@lingui/core/macro';
 import { Badge, Skeleton } from '@mantine/core';
 
-import { isTrue } from '@lib/functions/Conversion';
+import { isTrue } from '../functions/Conversion';
 
 export function PassFailButton({
   value,
