+++ src/frontend/src/components/importer/ImportDataSelector.tsx
@@ -10,6 +10,7 @@ import {
 import { type ReactNode, useCallback, useMemo, useState } from 'react';
 
 import { ActionButton } from '@lib/components/ActionButton';
+import { YesNoButton } from '@lib/components/YesNoButton';
 import { ApiEndpoints } from '@lib/enums/ApiEndpoints';
 import { apiUrl } from '@lib/functions/Api';
 import { cancelEvent } from '@lib/functions/Events';
@@ -29,7 +30,6 @@ import {
   RowDeleteAction,
   RowEditAction
 } from '../../tables/RowActions';
-import { YesNoButton } from '../buttons/YesNoButton';
 import { ProgressBar } from '../items/ProgressBar';
 import { RenderRemoteInstance } from '../render/Instance';
