+++ src/frontend/src/components/items/InfoItem.tsx
@@ -2,7 +2,7 @@ import { Trans } from '@lingui/react/macro';
 import { Code, Flex, Group, Text } from '@mantine/core';
 import { Link, type To } from 'react-router-dom';
 
-import { YesNoButton } from '../buttons/YesNoButton';
+import { YesNoButton } from '@lib/components/YesNoButton';
 import { DetailDrawerLink } from '../nav/DetailDrawer';
 
 export function InfoItem({
