+++ src/frontend/src/pages/Index/Settings/AdminCenter/ReportTemplatePanel.tsx
@@ -1,7 +1,7 @@
 import { ApiEndpoints } from '@lib/enums/ApiEndpoints';
 import { ModelType } from '@lib/enums/ModelType';
 import { t } from '@lingui/core/macro';
-import { YesNoButton } from '../../../../components/buttons/YesNoButton';
+import { YesNoButton } from '../../../../../lib/components/YesNoButton';
 import { TemplateTable } from '../../../../tables/settings/TemplateTable';
 
 function ReportTemplateTable() {
